Refactored modal directory.
