+++ src/modal/Modal.styles.ts
