+++ src/modal/Modal.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import styles from "./Modal.styles";
-import { getState, updateState } from "../../../state/State";
-import ILedgerWallet from "../../../interfaces/ILedgerWallet";
-import State from "../../../types/State";
+import { getState, updateState } from "../state/State";
+import ILedgerWallet from "../interfaces/ILedgerWallet";
+import State from "../types/State";
 
 declare global {
   // tslint:disable-next-line
